+++ tests/Mock/Polymorphic/Base.php
@@ -1,0 +1,18 @@
+<?php
+/*
+ * Copyright (c) Nate Brunette.
+ * Distributed under the MIT License (http://opensource.org/licenses/MIT)
+ */
+
+declare(strict_types=1);
+
+namespace Tebru\Gson\Test\Mock\Polymorphic;
+
+/**
+ * Class Base
+ *
+ * @author Nate Brunette <n@tebru.net>
+ */
+abstract class Base
+{
+}