+++ tests/Mock/Polymorphic/PolymorphicChild2.php
@@ -1,0 +1,17 @@
+<?php
+/*
+ * Copyright (c) Nate Brunette.
+ * Distributed under the MIT License (http://opensource.org/licenses/MIT)
+ */
+declare(strict_types=1);
+
+namespace Tebru\Gson\Test\Mock\Polymorphic;
+
+/**
+ * Class PolymorphicChild2
+ *
+ * @author Nate Brunette <n@tebru.net>
+ */
+class PolymorphicChild2 extends Base
+{
+}